Update semihost.rst
Please note that the RISC-V semihosting specification is in a "frozen" state. This means that no further changes or updates will be made to the document unless critical issues are identified during the public review cycle. Any future changes or extensions to the specification will be handled through follow-on extensions rather than updates to the frozen document.

For more detailed information on the RISC-V Semihosting Specification, refer to the official repository: